Maze Scene
I was not able to fix the magnitude of the coin spawn in the maze, but the randomized algo is already applied, just the position of the coins are not yet finished. Coin model is already there at UI assets. I created a script named coinGeneration in Coin folder under the scripts folder, it's not yet applied to anywhere as there's still a problem with the position of the coins, but you can just drag it in the CoinSpawn game object in hierarchy.